requirements.txt
absl-py==1.0.0
asgiref==3.5.2
astor==0.8.1
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
dill==0.2.7.1
Django==3.2.13
djangorestframework==3.13.1
fire==0.4.0
fonttools==4.29.1
gast==0.5.3
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
imageio==2.16.0
importlib-metadata==4.11.1
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.3.2
llvmlite==0.38.0
Markdown==3.3.6
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
msgpack==1.0.3
msgpack-numpy==0.4.7.1
networkx==2.6.3
numba==0.55.1
numpy @ file:///C:/ci_310/numpy_and_numpy_base_1643798589088/work
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
protobuf==3.19.4
psutil==5.9.0
pycocotools==2.0.4
pygame==2.1.2
pyparsing==3.0.7
pyreadline==2.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.2.0
pyzmq==22.3.0
requests==2.27.1
scikit-image==0.19.1
scikit-learn @ file:///C:/ci/scikit-learn_1642599122269/work
scipy @ file:///C:/ci/scipy_1641555141383/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
slidingwindow==0.0.14
sqlparse==0.4.2
tabulate==0.8.9
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorpack @ git+https://github.com/ppwwyyxx/tensorpack.git@fac024f0f72fd593ea243f0b599a51b11fe4effd
termcolor==1.1.0
tf-pose==0.11.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tqdm==4.62.3
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.13.3
zipp==3.7.0
